[ContainerRuntime] Separate runtimes from policies
Moves `ContainerRuntime` and the related subclasses out from
`sos/policies/__init__.py` into a new `sos/policies/runtimes` subdir.

Related: #2349
